requirements.txt
Django==5.2.7
psycopg2-binary==2.9.11
django-solo==2.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
sentry-sdk==1.40.0

